requirements.txt
google-cloud-storage==1.25.0
apache-beam==2.19.0
